Update dashboard to pull in real ordercloud data if configured to do so
If NEXT_PUBLIC_USE_REAL_DASHBOARD_DATA is set to true then the dashboard will show real OrderCloud data.

Note that the graph is not possible to implement without a backend component due to the sheer amount of API calls that would be required. If this type of analytics is required its recommended to compound that data on the backend separately and then simply retrieve the results client-side.

The data displayed has been modified slightly as well. Instead of monthly intervals we are showing weekly